Code compiles in C++, but does not boot :(.
This is the end of the line for me. I was expecting to make it work but no.
The original code in C, at least, boots. So I'll use it as it is as reference.
However I take with me many things like macros, function linking, or libraries. It's good to keep learning new stuff